Integration of array class mparray
The class mparray was introduced, but is not integrated in the other classes.
mparray defines an array implementation that is internally based upon Python lists. However, this fact is not visible to users of mparray. Thus, it would.be possible in the future to use other array implementations underneath without impacting users of mparray.